pyproject.toml
[tool.poetry]
name = "pydevmgr-serial"
version = "0.6.a0"
description = "based pydevmgr object for serial communication"
license = "CeCILL Free Software License Agreement v2.1"
authors = ["Sylvain Guieu <sylvain.guieu@univ-grenoble-alpes.fr>"]
readme = "README.md"
packages = [{include = "pydevmgr_serial"}]

[tool.poetry.dependencies]
python = "^3.7"
pydevmgr-core = ">=0.6.a1"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = [ "poetry-core"]
build-backend = "poetry.core.masonry.api"
